Add files via upload
This was our big ISC4U project where we (Morgan, Stephan, and I) used Java and intensive Object-Oriented-Programming concepts to complete a game similar to 2048 (except there are more modes where you can also choose '3' as the starter number). We wanted to make this program under the pretense that we were helping a friends' sister get better in math and problem solving! Within these compressed zip files, you'll find the entire project folder for the program and additionally, the JAR executable file!